Create Init Class
Clean up initial page. Make the code more readable.
